Update version to 1.1.3 and add TurnstileValidationService bean; refactor test structure
