+++ README.md
@@ -36,7 +36,7 @@ Add the following dependency to your `pom.xml`:
 <dependency>
     <groupId>com.digitalsanctuary</groupId>
     <artifactId>spring-cf-turnstile</artifactId>
-    <version>1.1.2</version>
+    <version>1.1.3</version>
 </dependency>
 ```
 
@@ -46,7 +46,7 @@ Add the following dependency to your `build.gradle`:
 
 ```groovy
 dependencies {
-    implementation 'com.digitalsanctuary:spring-cf-turnstile:1.1.2'
+    implementation 'com.digitalsanctuary:spring-cf-turnstile:1.1.3'
 }
 ```
