+++ build.gradle
@@ -13,7 +13,7 @@ import com.vanniktech.maven.publish.JavaLibrary
 import com.vanniktech.maven.publish.JavadocJar
 
 group 'com.digitalsanctuary.cf.turnstile'
-version '1.1.2'
+version '1.1.3'
 description = 'SpringBoot Cloudflare Turnstile Library'
 
 ext {
